Encapsulate Header for project consistency
